[WIP] DataファイルからＵＲＬを取得する処理をRepository層に移動
