+++ src/main/kotlin/com/demo/scraper/repository/ItemRepository.kt
@@ -1,0 +1,5 @@
+package com.demo.scraper.repository
+
+interface ItemRepository {
+    fun getUrl(code: String): String?
+}
