+++ src/main/kotlin/com/demo/scraper/service/ItemService.kt
@@ -1,19 +1,15 @@
 package com.demo.scraper.service
 
-import org.springframework.core.io.ClassPathResource
+import com.demo.scraper.repository.ItemRepository
+import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
-import java.io.File
 
 @Service
 class ItemService() {
-    // TODO: データファイルはRepository層に移動
-    private val file = "data/item.data"
-    private val dataFile: File = ClassPathResource(file).file
+    @Autowired
+    private lateinit var itemRepository: ItemRepository
 
     fun getUrl(code: String): String? {
-        return dataFile.readLines().stream()
-            .filter{ it.endsWith("/$code") }
-            .findFirst()
-            .orElse(null)
+        return itemRepository.getUrl(code)
     }
 }
